refac: Clean up logging packages
