+++ source/Logging/source/Logging.Tests/Logging.Tests.csproj
@@ -17,8 +17,9 @@ limitations under the License.
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <Nullable>enable</Nullable>
+        <RootNamespace>Energinet.DataHub.Core.Logging.Tests</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -38,7 +39,7 @@ limitations under the License.
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\RequestResponseMiddleware\RequestResponseMiddleware.csproj" />
+      <ProjectReference Include="..\Logging\Logging.csproj" />
     </ItemGroup>
 
 </Project>
