+++ source/Logging/source/Logging/Logging.csproj
@@ -17,17 +17,17 @@ limitations under the License.
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <LangVersion>9</LangVersion>
+    <TargetFramework>net7.0</TargetFramework>
+    <LangVersion>10</LangVersion>
     <Nullable>enable</Nullable>
-    <RootNamespace>Energinet.DataHub.Core.Logging.RequestResponseMiddleware</RootNamespace>
-    <AssemblyName>Energinet.DataHub.Core.Logging.RequestResponseMiddleware</AssemblyName>
+    <RootNamespace>Energinet.DataHub.Core.Logging</RootNamespace>
+    <AssemblyName>Energinet.DataHub.Core.Logging</AssemblyName>
   </PropertyGroup>
 
   <PropertyGroup>
     <PackageId>Energinet.DataHub.Core.Logging</PackageId>
-    <PackageVersion>2.3.1$(VersionSuffix)</PackageVersion>
-    <Title>Logging Middleware For Requests and Responses</Title>
+    <PackageVersion>3.0.0$(VersionSuffix)</PackageVersion>
+    <Title>Logging</Title>
     <Company>Energinet-DataHub</Company>
     <Authors>Energinet-DataHub</Authors>
     <PackageProjectUrl>https://github.com/Energinet-DataHub</PackageProjectUrl>
