+++ source/Logging/source/Logging/LoggingScope.cs
@@ -12,20 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Microsoft.Azure.Functions.Worker;
+using System.Collections.Concurrent;
 
-namespace RequestResponseMiddleware.Tests
+namespace Energinet.DataHub.Core.Logging;
+
+/// <summary>
+/// Logging scope intended to be used for any logging scope in the system for aligning logging.
+/// </summary>
+public class LoggingScope : ConcurrentDictionary<string, object>
 {
-    internal class FunctionBindingMetaData : BindingMetadata
-    {
-        public override string Type { get; }
-
-        public override BindingDirection Direction { get; }
-
-        public FunctionBindingMetaData(string type, BindingDirection direction)
-        {
-            Type = type;
-            Direction = direction;
-        }
-    }
 }
