--- source/Logging/source/LoggingScopeMiddleware/Internal/LoggingScope.cs
@@ -1,24 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System.Collections.Concurrent;
-
-namespace Energinet.DataHub.Core.Logging.LoggingScopeMiddleware.Internal;
-
-/// <summary>
-/// Logging scope intended to be used for any logging scope in the system for aligning logging.
-/// </summary>
-public class LoggingScope : ConcurrentDictionary<string, object>
-{
-}
