--- source/Logging/source/RequestResponseMiddleware.Tests/MockedFunctionInvocationFeatures.cs
@@ -1,52 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#nullable enable
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Microsoft.Azure.Functions.Worker;
-
-namespace RequestResponseMiddleware.Tests
-{
-    public class MockedFunctionInvocationFeatures : IInvocationFeatures
-    {
-        private readonly Dictionary<Type, object> _features = new();
-
-        public T? Get<T>()
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerator<KeyValuePair<Type, object>> GetEnumerator()
-        {
-            return _features.GetEnumerator();
-        }
-
-        public void Set<T>(T instance)
-        {
-            if (instance is null)
-            {
-                throw new ArgumentNullException(nameof(instance));
-            }
-
-            _features[typeof(T)] = instance;
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return _features.GetEnumerator();
-        }
-    }
-}
